software/raspberry_pi/ros2_ws/src/requirements.txt

board
adafruit-circuitpython-neopixel
adafruit-circuitpython-bme280
RPi.GPIO